requirements.txt
elevenlabs
python-dotenv
snowflake-connector-python
snowflake-sqlalchemy
sqlalchemy
sqlalchemy-utils
fastapi
google-genai
